test: extract test_tracks_call_float32_splice_plan to unit/tracks/
